PYTHON/Janmarg/janmarg/backend/requirements.txt
fastapi==0.75.0
uvicorn==0.17.0
torch==1.12.0
torchvision==0.13.0
transformers==4.21.1
pillow==9.0.1
exifread==2.3.2
requests==2.26.0
pydantic==1.9.0
pytest==7.1.2
pytest-cov==3.0.0